factorize some CLI tools
