+++ fabio/limaimage.py
@@ -33,7 +33,7 @@ __authors__ = ["Jérôme Kieffer"]
 __contact__ = "jerome.kieffer@esrf.fr"
 __license__ = "MIT"
 __copyright__ = "ESRF"
-__date__ = "15/06/2020" 
+__date__ = "23/04/2021"
 
 import logging
 logger = logging.getLogger(__name__)
@@ -60,12 +60,12 @@ class LimaImage(FabioImage):
         Set up initial values
         """
         if not h5py:
-            raise RuntimeError("fabio.EigerImage cannot be used without h5py. Please install h5py and restart")
+            raise RuntimeError("fabio.LimaImage cannot be used without h5py. Please install h5py and restart")
 
         FabioImage.__init__(self, data, header)
         self.dataset = [data]
         self.h5 = None
-        
+
     def __repr__(self):
         if self.h5 is not None:
             return "LImA-HDF5 dataset with %i frames from %s" % (self.nframes, self.h5.filename)
@@ -152,7 +152,6 @@ class LimaImage(FabioImage):
         if self.h5 is not None:
             self.h5.close()
             self.dataset = None
-
 
 
 # This is not compatibility with old code:
